requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
clang==5.0
click==8.0.1
cssselect==1.1.0
distlib==0.3.2
Django==3.2.6
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
fire==0.4.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==2.8
jieba3k==0.35.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
multitasking==0.0.9
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.0.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2017.4.5
requests==2.21.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tinysegmenter==0.3
tldextract==3.1.2
toposort==1.5
tqdm==4.31.1
typing-extensions==3.7.4.3
urllib3==1.24.3
virtualenv==20.6.0
virtualenv-clone==0.5.5
Werkzeug==2.0.1
wrapt==1.12.1
yfinance==0.1.63
